refactor(components): wrap all 3rd-party components
close #686
