+++ components/Headings/H1.tsx
@@ -1,0 +1,7 @@
+import Heading from './Heading';
+
+interface Props {}
+
+const H1 = (props: Props): JSX.Element => <Heading {...props} level={1} />;
+
+export default H1;
