+++ components/Headings/H2.tsx
@@ -1,0 +1,7 @@
+import Heading from './Heading';
+
+interface Props {}
+
+const H2 = (props: Props): JSX.Element => <Heading {...props} level={2} />;
+
+export default H2;
