+++ components/Headings/H3.tsx
@@ -1,0 +1,7 @@
+import Heading from './Heading';
+
+interface Props {}
+
+const H3 = (props: Props): JSX.Element => <Heading {...props} level={3} />;
+
+export default H3;
