+++ components/Headings/H4.tsx
@@ -1,0 +1,7 @@
+import Heading from './Heading';
+
+interface Props {}
+
+const H4 = (props: Props): JSX.Element => <Heading {...props} level={4} />;
+
+export default H4;
