+++ components/Headings/H6.tsx
@@ -1,0 +1,7 @@
+import Heading from './Heading';
+
+interface Props {}
+
+const H6 = (props: Props): JSX.Element => <Heading {...props} level={5} />;
+
+export default H6;
