+++ components/Headings/Heading.tsx
@@ -1,0 +1,8 @@
+import { Typography } from 'antd';
+import type { TitleProps } from 'antd/lib/typography/Title';
+
+const Heading = (props: TitleProps): JSX.Element => (
+  <Typography.Title {...props} />
+);
+
+export default Heading;
