+++ components/MDX/index.ts
@@ -1,7 +1,22 @@
-import Anchor from './Anchor';
-import Headings from './Headings';
-import Paragraph from './Paragraph';
-import Texts from './Texts';
+import Anchor from '@components/Anchor';
+import { H1, H2, H3, H4, H5, H6 } from '@components/Headings';
+import Paragraph from '@components/Paragraph';
+import { Delete, Emphasis, Strong } from '@components/Texts';
+
+const Headings = {
+  h1: H1,
+  h2: H2,
+  h3: H3,
+  h4: H4,
+  h5: H5,
+  h6: H6,
+};
+
+const Texts = {
+  strong: Strong,
+  em: Emphasis,
+  del: Delete,
+};
 
 const MDXComponents = {
   a: Anchor,
