+++ components/Texts/Emphasis.tsx
@@ -1,0 +1,7 @@
+import Text from './Text';
+
+interface Props {}
+
+const Emphasis = (props: Props): JSX.Element => <Text {...props} italic />;
+
+export default Emphasis;
