+++ components/Texts/Strong.tsx
@@ -1,0 +1,7 @@
+import Text from './Text';
+
+interface Props {}
+
+const Strong = (props: Props): JSX.Element => <Text {...props} strong />;
+
+export default Strong;
