+++ components/Texts/Text.tsx
@@ -1,0 +1,7 @@
+import { Typography } from 'antd';
+import type { TextProps } from 'antd/lib/typography/Text';
+
+
+const Text = (props: TextProps): JSX.Element => <Typography.Text {...props} />;
+
+export default Text;
