+++ components/Texts/index.ts
@@ -1,0 +1,3 @@
+export { default as Delete } from './Delete';
+export { default as Emphasis } from './Emphasis';
+export { default as Strong } from './Strong';
